+++ src/main/kotlin/com/jetbrains/rider/aspire/services/AspireResourceServiceViewDescriptor.kt
@@ -10,13 +10,27 @@ import com.intellij.ui.components.JBTabbedPane
 import com.jetbrains.rider.aspire.AspireBundle
 import com.jetbrains.rider.aspire.services.components.ResourceConsolePanel
 import com.jetbrains.rider.aspire.services.components.ResourceDashboardPanel
+import com.jetbrains.rider.aspire.settings.AspireSettings
 import com.jetbrains.rider.aspire.util.ASPIRE_RESOURCE
 import com.jetbrains.rider.aspire.util.getIcon
 import java.awt.BorderLayout
 import javax.swing.JPanel
 
-class AspireResourceServiceViewDescriptor(private val aspireResource: AspireResource) : ServiceViewDescriptor, DataProvider {
+class AspireResourceServiceViewDescriptor(private val aspireResource: AspireResource) : ServiceViewDescriptor,
+    DataProvider {
     private val toolbarActions = ActionManager.getInstance().getAction("Aspire.Resource") as ActionGroup
+
+    private val tabs = JBTabbedPane().apply {
+        addTab(AspireBundle.message("service.tab.dashboard"), ResourceDashboardPanel(aspireResource))
+        addTab(AspireBundle.message("service.tab.console"), ResourceConsolePanel(aspireResource))
+        if (AspireSettings.getInstance().openConsoleView) {
+            selectedIndex = 1
+        }
+    }
+
+    private val panel = JPanel(BorderLayout()).apply {
+        add(tabs, BorderLayout.CENTER)
+    }
 
     override fun getPresentation() = PresentationData().apply {
         val icon = getIcon(aspireResource)
@@ -25,14 +39,7 @@ class AspireResourceServiceViewDescriptor(private val aspireResource: AspireReso
     }
 
     override fun getContentComponent(): JPanel {
-        val tabs = JBTabbedPane()
-        tabs.addTab(AspireBundle.message("service.tab.dashboard"), ResourceDashboardPanel(aspireResource))
-        tabs.addTab(AspireBundle.message("service.tab.console"), ResourceConsolePanel(aspireResource))
-
-        val panel = JPanel(BorderLayout()).apply {
-            add(tabs, BorderLayout.CENTER)
-        }
-
+        tabs.setComponentAt(0, ResourceDashboardPanel(aspireResource))
         return panel
     }